refactor: 데이터 매핑 기능 ProfileMapper로 파일 분리
