+++ src/apis/profile/profileApi.js
@@ -1,19 +1,5 @@
-import { educationLevel, educationStatus, jobData } from '../../data/profileData';
 import apiClient from '../axiosInstance';
-
-const transformProfileData = (profile) => ({
-    ...profile,
-    educationLevel: educationLevel.find((item) => item.label === profile.educationLevel)?.value || '',
-    educationStatus: educationStatus.find((item) => item.label === profile.educationStatus)?.value || '',
-    job: jobData.find((item) => item.label === profile.job)?.id || '',
-});
-
-const transformProfileResponse = (responseData) => ({
-    ...responseData,
-    educationLevel: educationLevel.find((item) => item.value === responseData.educationLevel)?.label || '',
-    educationStatus: educationStatus.find((item) => item.value === responseData.educationStatus)?.label || '',
-    job: responseData.job.name,
-});
+import { transformProfileData, transformProfileResponse } from '../../utils/Profile/ProfileMapper';
 
 export const saveProfile = async (profile) => {
     const transformData = transformProfileData(profile);
