+++ src/data/profileData.js
@@ -15,26 +15,3 @@ export const profileEmptyData = {
     educationStatus: '',
     job: '',
 };
-
-export const educationLevel = [
-    { label: '중학교 이하', value: 'MIDDLE' },
-    { label: '고등학교', value: 'HIGH' },
-    { label: '전문대학', value: 'JUNIOR_COLLEGE' },
-    { label: '대학교', value: 'UNIVERSITY' },
-    { label: '석사', value: 'MASTER' },
-    { label: '박사', value: 'DOCTOR' },
-];
-
-export const educationStatus = [
-    { label: '재학', value: 'ENROLLED' },
-    { label: '휴학', value: 'ON_LEAVE' },
-    { label: '졸업', value: 'GRADUATED' },
-    { label: '수료', value: 'COMPLETED' },
-];
-
-export const jobData = [
-    { label: '프론트엔드 개발자', id: 1 },
-    { label: '백엔드 개발자', id: 2 },
-    { label: 'PM(Product/Project Manager)', id: 3 },
-    { label: 'Designer', id: 4 },
-];
